% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taster_notes.R
\name{experiment_df}
\alias{experiment_df}
\title{make a dataframe for experiments}
\usage{
experiment_df(
  filepath = file.path(config::get("tastermonial")$datadir,
    "tastermonial_experiments.csv")
)
}
\description{
Currently just reads the experiments as written to a CSV
Ideally this lets you generate a new CSV as appropriate
Needs to be filled in with code that reads a list of
experiments and prepares them for writing to the database
}
