% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glucose.R
\name{fill_glucose_records_from_scratch}
\alias{fill_glucose_records_from_scratch}
\title{Read all CSV files again and enter them into the database}
\usage{
fill_glucose_records_from_scratch(con, drop = TRUE)
}
\value{
dataframe
}
\description{
Read all CSV files again and enter them into the database
}
