% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taster_notes.R
\name{fill_taster_notes_from_scratch}
\alias{fill_taster_notes_from_scratch}
\title{Fill all Tastermonial notes}
\usage{
fill_taster_notes_from_scratch(con, taster_notes_df, drop = TRUE)
}
\arguments{
\item{con}{database connection}

\item{drop}{nuke current table and start over if \code{TRUE} (default)}

\item{taster_note_df}{a notes dataframe}
}
\description{
Fill all Tastermonial notes
}
