% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taster_notes.R
\name{id_from_initial}
\alias{id_from_initial}
\title{find user_id from first few chars in a name}
\usage{
id_from_initial(name_initials)
}
\description{
Tastermonial users are kept in a JSON object with $email $name and $uid fields.
Unfortunately, the names are often not consistent with what we find in Libreview (e.g. Bude is Buda).
so this function attempts to match people based on the first name in the libreview list that matches initials.
}
