% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taster_notes.R
\name{id_from_taster}
\alias{id_from_taster}
\title{Find user_id from a Tastermonial Json object}
\usage{
id_from_taster(json_object)
}
\arguments{
\item{json_object}{a JSON-formatted object}
}
\description{
Find user_id from a Tastermonial Json object
}
