% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load.R
\name{load_db_efficient}
\alias{load_db_efficient}
\title{Create a taster_db object and load all available Libreview glucose files from scratch}
\usage{
load_db_efficient(db_name = "sqldb")
}
\arguments{
\item{db_name}{name of Postgres SQL catalog (aka database)}
}
\description{
Create a taster_db object and load all available Libreview glucose files from scratch
}
