% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load.R
\name{load_db_extra}
\alias{load_db_extra}
\title{load extra items into database}
\usage{
load_db_extra(con)
}
\arguments{
\item{con}{database connection}
}
\description{
This is a hack function that loads the database with items that
for whatever reason weren't loaded in the main \code{load_db()} function.
Now it loads notes from one Libreview CSV file.
}
