% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glucose.R
\name{load_libreview_csv_from_directory}
\alias{load_libreview_csv_from_directory}
\title{Unified dataframe for all glucose CSV files in \code{path}}
\usage{
load_libreview_csv_from_directory(path = config::get("tastermonial")$datadir)
}
\arguments{
\item{path}{file path to a directory of libreview CSV files.}
}
\value{
dataframe including \code{user_id} matching those for data in the notes_records
}
\description{
Read glucose files in \code{path} and return one big dataframe with all glucose values.
As a neat trick, look up the \code{user_id} based on entries in \code{user_df_from_libreview} and
add that to the dataframe.
Assumes it's a valid file if it has the string "glucose" in its name.
}
