% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{name_for_user_id}
\alias{name_for_user_id}
\title{Find username associated with an ID}
\usage{
name_for_user_id(user_id)
}
\arguments{
\item{user_id}{user ID}
}
\value{
character string of the username for that ID
}
\description{
Find username associated with an ID
}
