% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taster_notes.R
\name{notes_df_from_taster2111}
\alias{notes_df_from_taster2111}
\title{Notes Dataframe From Specified User (2111 Format)}
\usage{
notes_df_from_taster2111(
  filepath = file.path(config::get("tastermonial")$datadir, "table-data-new.csv"),
  username = "Richard Sprague",
  user_id = -1
)
}
\arguments{
\item{filepath}{path to a CSV file in 2111 format}

\item{username}{char string full name for user in CSV file}
}
\description{
Notes Dataframe From Specified User (2111 Format)
}
