% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load.R
\name{psi_fill_database_from_scratch}
\alias{psi_fill_database_from_scratch}
\title{Fill database after dropping}
\usage{
psi_fill_database_from_scratch(
  conn_args = config::get("dataconnection"),
  drop = TRUE
)
}
\description{
For debugging and dev purposes only. Loads the database tables from scratch.
@noRd
}
