% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glucose.R
\name{psi_fill_glucose_records_from_scratch}
\alias{psi_fill_glucose_records_from_scratch}
\title{Read all CSV files again and enter them into the database}
\usage{
psi_fill_glucose_records_from_scratch(
  conn_args = config::get("dataconnection"),
  drop = TRUE
)
}
\value{
dataframe
}
\description{
Read all CSV files again and enter them into the database
}
