% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load.R
\name{psi_fill_notes_records_from_scratch}
\alias{psi_fill_notes_records_from_scratch}
\title{Nuke all notes records and start over}
\usage{
psi_fill_notes_records_from_scratch(
  conn_args = config::get("dataconnection"),
  drop = TRUE
)
}
\description{
Nuke all notes records and start over
}
