% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{psi_list_objects}
\alias{psi_list_objects}
\title{List all objects in the current PSI database}
\usage{
psi_list_objects(conn_args = config::get("dataconnection"))
}
\description{
List all objects in the current PSI database
}
