% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{psi_make_database_if_necessary}
\alias{psi_make_database_if_necessary}
\title{Make new Postgres database if one doesn't already exist.}
\usage{
psi_make_database_if_necessary(conn_args = config::get("dataconnection"))
}
\arguments{
\item{conn_args}{a valid database connection (assumes Postgres)}
}
\description{
Make new Postgres database if one doesn't already exist.
}
