% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{psi_make_table_if_necessary}
\alias{psi_make_table_if_necessary}
\title{Make new database tables if necessary}
\usage{
psi_make_table_if_necessary(
  conn_args = config::get("dataconnection"),
  table_name = "glucose_records",
  table = GLUCOSE_DATA_FRAME
)
}
\arguments{
\item{conn_args}{connection}

\item{table_name}{character string name for the table.}

\item{table}{a valid glucose data frame. Never use the default value unless you are testing.}
}
\value{
NULL if table already exists. Otherwise creates the table and returns TRUE invisibly.
}
\description{
Make new database tables if necessary
}
