% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{psi_make_table_with_index}
\alias{psi_make_table_with_index}
\title{Make a new database tables with \code{index}}
\usage{
psi_make_table_with_index(
  conn_args = config::get("dataconnection"),
  table_name = "experiments",
  table = NULL,
  index = NULL
)
}
\arguments{
\item{conn_args}{connection}

\item{table_name}{character string name for the table.}

\item{table}{a valid glucose data frame. Never use the default value unless you are testing.}

\item{index}{(list) table column to be used for index}
}
\value{
NULL if table already exists. Otherwise creates the table and returns TRUE invisibly.
}
\description{
Make a new database tables with \code{index}
}
