% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{psi_table_df}
\alias{psi_table_df}
\title{Show table as a dataframe}
\usage{
psi_table_df(
  conn_args = config::get("dataconnection"),
  table_name = "glucose_records"
)
}
\arguments{
\item{conn_args}{valid data connection}

\item{table_name}{character string of the table name (default: \code{glucose_records})}
}
\value{
dataframe
}
\description{
Read contents of \code{table} from the database and return it as a dataframe
}
