% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load.R
\name{psi_write_glucose}
\alias{psi_write_glucose}
\title{Write a glucose dataframe to the database}
\usage{
psi_write_glucose(
  conn_args = config::get("dataconnection"),
  user_id = 1235,
  new_table = glucose_df_from_libreview_csv(user_id = 1235)
)
}
\arguments{
\item{new_table}{valid formatted glucose dataframe}
}
\description{
WARNING: always delete the table before running this on a \code{user_id} that's already in the database.
(it's not finished yet and doesn't take into account previous entries)
}
