% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load.R
\name{psi_write_notes}
\alias{psi_write_notes}
\title{Write a Notes CSV to the notes table in the database}
\usage{
psi_write_notes(
  conn_args = config::get("dataconnection"),
  user_id = 1235,
  new_table = notes_df_from_csv(user_id = 1235),
  dry_run = TRUE
)
}
\arguments{
\item{user_id}{user ID}

\item{new_table}{valid formatted notes dataframe}

\item{dry_run}{(default = TRUE). Run without actually writing to the database}
}
\description{
WARNING: this may not work correctly. debug before using
psi_write_notes(user_id = 1234, new_table = notes_df_from_glucose_table(user_id = 1234))
}
