% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taster_notes.R
\name{taster_classify_food}
\alias{taster_classify_food}
\title{Classify a Tastermonial food into limited categories}
\usage{
taster_classify_food(foodname)
}
\arguments{
\item{foodname}{a string representation of a name}
}
\value{
character string representing simplifed name
}
\description{
Data from the Tastermonial Retool database (see \code{taster_from_retool_csv()}) is
not standardized and can often be quite messy.  Before converting it to the \code{notes_records} format,
this function will apply the transformation from \verb{Tastermonial Name Mapping.csv} to
return a canonical name for each food type.
}
