% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taster_notes.R
\name{transform_old_typeform}
\alias{transform_old_typeform}
\title{(one-time only) Make a version of the Typeform data with correct timezone info}
\usage{
transform_old_typeform()
}
\value{
dataframe
}
\description{
This function should only be needed one time. Use it to write a new file
and have that be the source for future access to the old Typeform data.
}
