% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{user_df_from_csv}
\alias{user_df_from_csv}
\title{Return user list from Tastermonial Libreview download}
\usage{
user_df_from_csv(
  file = file.path(config::get("tastermonial")$datadir,
    "Tastermonial_allPatients_dashboard.csv")
)
}
\arguments{
\item{file}{the main file downloaded from a Libreview practice ID}
}
\description{
A Libreview "practice" stores all its user information in a single
CSV file, which this function will convert into a canonical dataframe.
}
