% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{username_for_id}
\alias{username_for_id}
\title{Find username associated with an ID}
\usage{
username_for_id(user_id)
}
\arguments{
\item{user_id}{user ID}
}
\value{
character string of the username for that ID
}
\description{
Find username associated with an ID
}
