% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glucose.R
\name{write_glucose}
\alias{write_glucose}
\title{Write a glucose dataframe to the database}
\usage{
write_glucose(new_table)
}
\arguments{
\item{new_table}{valid formatted glucose dataframe}
}
\description{
(it's not finished yet and doesn't take into account previous entries)
}
