% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{write_user_list_from_scratch}
\alias{write_user_list_from_scratch}
\title{writes an entire user list to the database}
\usage{
write_user_list_from_scratch(con, user_list = user_df_from_db(), drop = TRUE)
}
\arguments{
\item{user_list}{valid user list as a dataframe}
}
\description{
writes an entire user list to the database
}
