% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load.R
\name{load_db}
\alias{load_db}
\title{Crate a taster_db object and load all available Libreview glucose files from scratch}
\usage{
load_db(ps_database = "sandbox")
}
\arguments{
\item{ps_database}{name of Postgres SQL catalog (aka database)}
}
\description{
Crate a taster_db object and load all available Libreview glucose files from scratch
}
