% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taster_notes.R
\name{taster_classify_food}
\alias{taster_classify_food}
\title{Classify a Tastermonial food into limited categories}
\usage{
taster_classify_food(foodname)
}
\arguments{
\item{foodname}{a string representation of a name}
}
\value{
character string representing simplifed name
a CSV file with columns \code{pid}, \code{names}, and \code{simpleName} to convert from each format
}
\description{
Classify a Tastermonial food into limited categories
}
