% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taster_notes.R
\name{taster_from_retool_csv}
\alias{taster_from_retool_csv}
\title{Read notes from the iPhone app csv}
\usage{
taster_from_retool_csv()
}
\value{
dataframe
}
\description{
There's no need to correct for time zone because the Retool data is already
timestamped with UTC.
}
