% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taster_notes.R
\name{taster_raw}
\alias{taster_raw}
\title{Read a raw Tastermonial CGM log (as downloaded from retool)}
\usage{
taster_raw(
  filepath = file.path(config::get("tastermonial")$datadir, "table-data.csv")
)
}
\arguments{
\item{filepath}{path to a CSV download}
}
\description{
Read a raw Tastermonial CGM log (as downloaded from retool)
}
