% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\docType{data}
\name{user_df_from_libreview}
\alias{user_df_from_libreview}
\title{Users known to Libreview Practice Portal}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26 rows and 6 columns.
}
\usage{
user_df_from_libreview
}
\description{
A dataframe of all users and their ids, taken from the Libreview practice portal
}
\keyword{datasets}
