% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{user_df_from_libreview}
\alias{user_df_from_libreview}
\title{Users known to Libreview Practice Portal}
\usage{
user_df_from_libreview()
}
\value{
A dataframe of all users and their ids, taken from the Libreview practice portal
}
\description{
Libreview Practice Portal offers a complete list of the names
of users who have submitted glucose reports.
}
